requirements.txt
django==4.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-cors-headers==4.0.0
django-filter==2.4.0
djoser==2.1.0
Pillow==11.1.0